+++ util/messages.h
@@ -49,35 +49,35 @@ void sysdie(const char *, ...)
  * of those handlers.  These functions are not thread-safe; they set global
  * variables.
  */
-void message_handlers_debug(int count, ...);
-void message_handlers_notice(int count, ...);
-void message_handlers_warn(int count, ...);
-void message_handlers_die(int count, ...);
+void message_handlers_debug(unsigned int count, ...);
+void message_handlers_notice(unsigned int count, ...);
+void message_handlers_warn(unsigned int count, ...);
+void message_handlers_die(unsigned int count, ...);
 
 /*
  * Some useful handlers, intended to be passed to message_handlers_*.  All
  * handlers take the length of the formatted message, the format, a variadic
  * argument list, and the errno setting if any.
  */
-void message_log_stdout(int, const char *, va_list, int)
+void message_log_stdout(size_t, const char *, va_list, int)
     __attribute((__nonnull__));
-void message_log_stderr(int, const char *, va_list, int)
+void message_log_stderr(size_t, const char *, va_list, int)
     __attribute((__nonnull__));
-void message_log_syslog_debug(int, const char *, va_list, int)
+void message_log_syslog_debug(size_t, const char *, va_list, int)
     __attribute((__nonnull__));
-void message_log_syslog_info(int, const char *, va_list, int)
+void message_log_syslog_info(size_t, const char *, va_list, int)
     __attribute((__nonnull__));
-void message_log_syslog_notice(int, const char *, va_list, int)
+void message_log_syslog_notice(size_t, const char *, va_list, int)
     __attribute((__nonnull__));
-void message_log_syslog_warning(int, const char *, va_list, int)
+void message_log_syslog_warning(size_t, const char *, va_list, int)
     __attribute((__nonnull__));
-void message_log_syslog_err(int, const char *, va_list, int)
+void message_log_syslog_err(size_t, const char *, va_list, int)
     __attribute((__nonnull__));
-void message_log_syslog_crit(int, const char *, va_list, int)
+void message_log_syslog_crit(size_t, const char *, va_list, int)
     __attribute((__nonnull__));
 
 /* The type of a message handler. */
-typedef void (*message_handler_func)(int, const char *, va_list, int);
+typedef void (*message_handler_func)(size_t, const char *, va_list, int);
 
 /* If non-NULL, called before exit and its return value passed to exit. */
 extern int (*message_fatal_cleanup)(void);
